refactor to do list to use another component
